+++ src/main/java/step3/racingcar/domain/CarFactory.java
@@ -1,4 +1,4 @@
-package step3.racingcar;
+package step3.racingcar.domain;
 
 import java.util.Arrays;
 import java.util.List;
