+++ src/main/java/step3/racingcar/domain/CarGroups.java
@@ -1,4 +1,4 @@
-package step3.racingcar;
+package step3.racingcar.domain;
 
 import java.util.Comparator;
 import java.util.List;
