+++ src/main/java/step3/racingcar/domain/Message.java
@@ -1,4 +1,4 @@
-package step3.racingcar;
+package step3.racingcar.domain;
 
 public class Message {
