+++ src/main/java/step3/racingcar/domain/MovingStrategy.java
@@ -1,4 +1,4 @@
-package step3.racingcar;
+package step3.racingcar.domain;
 
 public interface MovingStrategy {
