+++ src/test/java/step3/racingcar/CarGroupsTest.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import step3.racingcar.domain.CarFactory;
+import step3.racingcar.domain.CarGroups;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Predicate;
