+++ src/test/java/step3/racingcar/CarTest.java
@@ -5,6 +5,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import step3.racingcar.domain.Car;
+import step3.racingcar.domain.CarFactory;
+import step3.racingcar.domain.RandomMovingStrategy;
+
 import java.util.List;
 import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
