+++ src/test/java/step3/racingcar/RacingGameTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import step3.racingcar.controller.RacingGame;
+
 import java.util.Arrays;
 import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
